backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.9"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.123.10",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
    "greenlet>=3.2.4 ; sys_platform == 'darwin'",
    "pyjwt>=2.10.1",
    "bcrypt>=5.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
